--- systest/mutations-and-queries/helper_test.go
@@ -1,42 +1,0 @@
-//go:build integration || upgrade
-
-/*
- * Copyright 2017-2023 Dgraph Labs, Inc. and Contributors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-//nolint:lll
-package main
-
-import (
-	"context"
-
-	"github.com/pkg/errors"
-
-	"github.com/dgraph-io/dgraph/dgraphtest"
-	"github.com/dgraph-io/dgraph/x"
-)
-
-func doGrpcLogin(ssuite *SystestTestSuite) (*dgraphtest.GrpcClient, func(), error) {
-	gcli, cleanup, err := ssuite.dc.Client()
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "error creating grpc client")
-	}
-	err = gcli.LoginIntoNamespace(context.Background(),
-		dgraphtest.DefaultUser, dgraphtest.DefaultPassword, x.GalaxyNamespace)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "groot login into galaxy namespace failed")
-	}
-	return gcli, cleanup, nil
-}
