+++ systest/mutations-and-queries/integration_test.go
@@ -19,14 +19,12 @@
 package main
 
 import (
-	"context"
 	"testing"
 
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 
 	"github.com/dgraph-io/dgraph/dgraphtest"
-	"github.com/dgraph-io/dgraph/x"
 )
 
 type SystestTestSuite struct {
@@ -40,11 +38,9 @@ func (ssuite *SystestTestSuite) SetupTest() {
 
 func (ssuite *SystestTestSuite) SetupSubTest() {
 	t := ssuite.T()
-	gcli, cleanup, err := ssuite.dc.Client()
+	gcli, cleanup, err := doGrpcLogin(ssuite)
 	defer cleanup()
 	require.NoError(t, err)
-	require.NoError(t, gcli.LoginIntoNamespace(context.Background(),
-		dgraphtest.DefaultUser, dgraphtest.DefaultPassword, x.GalaxyNamespace))
 	require.NoError(t, gcli.DropAll())
 }
